advanced-rag/pyproject.toml
[tool.poetry]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = ["adamscarlat <ascarlat@vividcloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
langchain = "^0.2.7"
langgraph = "^0.1.8"
langchainhub = "^0.1.20"
langchain-community = "^0.2.7"
tavily-python = "^0.3.4"
langchain-chroma = "^0.1.2"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
langchain-openai = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
